Assignment 2/requirements.txt
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-sql
azure-mgmt-storage
azure-identity
